requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
aiofiles==23.2.1
pandas==2.2.2
pyarrow==16.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.3
lxml==5.2.1
kaggle==1.6.14
sqlalchemy==2.0.31
aiosqlite==0.20.0
jinja2==3.1.4
python-multipart==0.0.9
sse-starlette==2.1.0
python-dotenv==1.0.1
